+++ project/components/sections/settings/profile-settings-form.tsx
@@ -1,6 +1,6 @@
-import Select from "@/components/select";
 import Button from "@/components/ui/buttons/button";
 import Input from "@/components/ui/input-fields/input";
+import Select from "@/components/ui/input-fields/select/select";
 import { cn } from "@/lib/utils";
 
 interface ProfileSettingsFormProps {
